[refactor] sample内部を整理
