+++ sample/src/main/kotlin/dataset/mnist/MnistUtils.kt
@@ -1,0 +1,49 @@
+package dataset.mnist
+
+import com.wsr.NetworkBuilder
+import com.wsr.common.IOType
+import com.wsr.layers.affine.affine
+import com.wsr.layers.bias.bias
+import com.wsr.layers.conv.convD1
+import com.wsr.layers.function.relu.relu
+import com.wsr.layers.function.softmax.softmax
+import com.wsr.layers.pool.maxPool
+import maxIndex
+import java.util.Random
+
+fun createMnistModel(epoc: Int, seed: Int? = null) {
+    val network = NetworkBuilder.inputD2(x = 28, y = 28, rate = 0.01, seed = seed)
+        .convD1(filter = 30, kernel = 5, stride = 1, padding = 0).bias().relu().maxPool(2)
+        .reshapeD1()
+        .affine(neuron = 512).bias().relu()
+        .affine(neuron = 10).softmax()
+        .build()
+
+    val random = seed?.let { Random(seed.toLong()) } ?: Random()
+
+    val dataset = MnistDataset.read().shuffled(random)
+    val (train, test) = dataset.take(50000) to dataset.takeLast(10000).take(100)
+    println("${dataset.size}")
+
+    (1..epoc).forEach { epoc ->
+        println("epoc: $epoc")
+        train.shuffled(random).take(230).forEachIndexed { i, data ->
+            network.train(
+                input = IOType.D2(
+                    data.pixels.toMutableList(),
+                    listOf(28, 28),
+                ),
+                label = IOType.D1(10) { if (data.label == it) 1.0 else 0.0 },
+            )
+            if (i % 100 == 0) println("trained: $i")
+        }
+    }
+    test.count { data ->
+        network.expect(
+            input = IOType.D2(
+                data.pixels.toMutableList(),
+                listOf(28, 28),
+            ),
+        ).value.maxIndex() == data.label
+    }.let { println(it.toDouble() / test.size.toDouble()) }
+}
